+++ kernel/vmm.c
@@ -2,8 +2,8 @@
 #include "param.h"
 #include "riscv.h"
 #include "spinlock.h"
-#include "proc.h"
 #include "hashmap.h"
+#include "proc.h"
 #include "hashmapPage.h"
 #include "vmm.h"
 #include "defs.h"
@@ -38,9 +38,9 @@ void randomSampling() {
   release(&vmmState.samplingTimeLock);
 }
 
-void printRegisteredVM(int key, void *value) {
+void printRegisteredVM(uint64 key, void *value) {
   struct proc *p = (struct proc *) value;
-  printf("PID: %d --> VM: (%d->%d, %p)\n", key, p->parent->pid, p->pid, p);
+  printf("PID: %d --> VM: (%d->%d, %p)\n", (int) key, p->parent->pid, p->pid, p);
 }
 
 void printRegisteredVMs() {
